requirements.txt
flask>=3.0,<4.0
azure-storage-blob>=12.20,<13.0
gunicorn>=22.0,<23.0
python-dotenv>=1.0,<2.0
celery[redis]>=5.3,<6.0
redis>=5.0,<6.0
